create a model based autocomplete component in order to allow avoiding serialising the choices.
